+++ json/albums.json
@@ -1,1 +1,1 @@
-[{"album_name":"Wow","publish_date":"2021-10-12","on_spotify?":false,"genre":"Rap"},{"album_name":"Hello","publish_date":"4/9/2021","on_spotify?":false,"genre":"HipHop"}]
+[{"album_name":"Wow","publish_date":"2021-10-12","on_spotify?":false,"genre":"Rap"},{"album_name":"Hello","publish_date":"4/9/2021","on_spotify?":false,"genre":"HipHop"},{"album_name":"hh","publish_date":"nn","on_spotify?":false,"genre":"bb"},{"album_name":"ff","publish_date":"ff","on_spotify?":false,"genre":"ff"},{"album_name":"fff","publish_date":"bb","on_spotify?":false,"genre":"dd"}]
